Expose parametric and curve type variables
Exposes the "value" property for both parametric
and generic curve types. The value will be:

curveType, one of:
  None, in the case of "Identity" type curve
  Float, in the case of "Power Function" type curve
  Array of Floats, in the case of 1D type curve

parametricCurveType will be the function as a string

In addition to the two basic value properties exposed,
curveType exposes the curvetype property, which will
be one of "Identity", "Power Function", or "1D Curve"
string.

For parametricCurveType, the parameters property is
exposed, which will be a dictionary of the possible
parameters, along with their float values.